PB1033+PB1058+PB1178: one saturating narrowing for integer values
Train 60, cluster w59b (implementer branch worktree-agent-a488d53c4b7e35173,
head b2bd13d08, base df452130e).

Mechanism: emit-time (int)/(long) casts on COBOL integer values wrapped
past the host carrier, and binder int.Parse/TryParse sites either crashed
(OverflowException) or silently dropped the value. Measured before the fix:
GO TO DEPENDING 4294967297 went to P1 (ISO 14.9.17.4 GR2: no transfer
occurs); ref-mod start 4294967297 read position 1 (8.4.3.3.4 5c);
STOP RUN WITH ERROR STATUS 4294967296 exited 0; literal ADVANCING/RETRY/
TIMES counts failed Roslyn (CS0221/CS1021); PAGE LIMIT / COLUMN / LINAGE
77777777777 crashed the compiler; OCCURS and RECORD/BLOCK CONTAINS
77777777777 compiled with no diagnostic.

Fix shape: ONE runtime narrowing (CobolNum.Position, new int twin
Position32 + IntegerLiteralValue) reached only through RuntimeApi.HostInt32/
HostInt64 (literal variants fold at compile time with the same functions);
ONE binder reader IntegerOperandRules.HostValue (saturating, never throws;
16 parse sites rerouted, LineInteger deleted) behind ONE screen,
IntegerOperandPass -> COBOLNET2427 (ISO 4.5 implementation limit), with
FullValueSlots (PERFORM TIMES, WRITE ADVANCING, DYNAMIC LENGTH LIMIT)
exempt. Arms: GO TO DEPENDING, ref-mod start/length (SpecifiedRefModLength
keeps a saturated length off the OMITTED sentinel), ACCEPT ref-mod width,
PERFORM TIMES, ADVANCING, LINAGE, RETRY TIMES/SECONDS (retryAmount long end
to end), STRING/UNSTRING pointers, UNSTRING TALLYING (Int128), RELATIVE KEY,
CONTINUE AFTER, B-SHIFT, ODO current-count reads, exit status (emitter +
RunUnit.ExitStatus clamps). The duplicated, overflowing runtime ref-mod
range test is merged into one overflow-safe RefModOutOfRange.
Drift: semgrep cobolnet-integer-value-cast-narrowed and
cobolnet-integer-n-parsed-outside-host-value; IntegerOperandSlotDriftTests.

Goldens: 2002/pb1033_integer_operand_full_value; negative
pb1058-integer-operand-beyond-limit (COBOLNET2427, reject-at 85 2002 2014
2023); StopGobackExitCodeTests.StatusBeyondTheHostExitCode_IsClampedNeverWrapped.
Characterization snapshots char_occurs/char_perform/char_string_ops
re-baselined (intended diffs).

Register: PB1033 landed (closes GR-14.9.17.4-2, GR-14.7.9.3-1,
GR-14.9.43.4-4); PB1058 landed (closes_rows [] with reason); PB1178 stays
open (closes GR-14.9.42.4-5; the ERROR-vs-value question is answered by the
orchestrator decision already on the note, both paragraphs kept on the
rebase); PB1163 gains GR-14.7.9.3-4. Verdict batch w59b re-applied on the
merged tree with record_verdicts.py: 8 rows, GAP 1458 -> 1454.

+++ src/Cobol.Net.Runtime/IO/CobolFile.cs
@@ -237,7 +237,7 @@ public static class CobolFile
     /// carries the independent per-file-name TAPE PHRASE, which a sharing-phrase OPEN may also write
     /// (§14.9.27.2 prints both in one general format).</summary>
     public static void OpenShared(string name, FileOpenMode mode, bool hasSharingOverride, FileSharing sharingOverride,
-        FileRetryKind retryKind, int retryAmount, OpenTapePhrase tape, string assign, bool assignDynamic, LinagePage? page)
+        FileRetryKind retryKind, long retryAmount, OpenTapePhrase tape, string assign, bool assignDynamic, LinagePage? page)
         => _reg.OpenShared(name, mode, hasSharingOverride, sharingOverride, retryKind, retryAmount, tape,
             assign, assignDynamic, page);
 
@@ -247,7 +247,7 @@ public static class CobolFile
     /// <paramref name="ignoringLock"/> the INDEPENDENT IGNORING LOCK phrase (§14.9.30.2's other bracket, GR12).
     /// A Format-1 read of any organization uses <see cref="ReadShared"/> instead.</summary>
     public static string ReadKeyedShared(string name, int keyIndex, string keyedRecordImage, FileRecordLock phrase,
-        bool ignoringLock, FileRetryKind retryKind, int retryAmount, out string image)
+        bool ignoringLock, FileRetryKind retryKind, long retryAmount, out string image)
         => _reg.ReadKeyedShared(name, keyIndex, keyedRecordImage, phrase, ignoringLock, retryKind, retryAmount, out image);
 
     /// <summary>The ONE governed FORMAT-1 READ — sequential, relative and indexed (§9.1.16 / §14.9.30.4 GR9–GR12
@@ -256,7 +256,7 @@ public static class CobolFile
     /// <paramref name="advancingOnLock"/> and <paramref name="ignoringLock"/> are two alternatives of the SAME
     /// printed bracket, so at most one is ever true; <paramref name="phrase"/> is the other bracket.</summary>
     public static string ReadShared(string name, bool previous, FileRecordLock phrase, bool advancingOnLock,
-        bool ignoringLock, FileRetryKind retryKind, int retryAmount, out string image)
+        bool ignoringLock, FileRetryKind retryKind, long retryAmount, out string image)
         => _reg.ReadShared(name, previous, phrase, advancingOnLock, ignoringLock, retryKind, retryAmount, out image);
 
     /// <summary>⛔ THE ONE WRITE ENTRY THE EMITTER RENDERS, every organization and every print-control shape
@@ -264,21 +264,21 @@ public static class CobolFile
     /// see <see cref="WriteAdvanceKind"/> for why a WRITE's presentation shape may not pick its own entry
     /// (kb/Work PB683).</summary>
     public static string WriteShared(string name, string image, int length, FileRecordLock phrase,
-        FileRetryKind retryKind, int retryAmount, LinagePage? page, WriteAdvance advance = default)
+        FileRetryKind retryKind, long retryAmount, LinagePage? page, WriteAdvance advance = default)
         => _reg.WriteShared(name, image, length, phrase, retryKind, retryAmount, page, advance);
 
     /// <summary>Governed REWRITE for a sharing-active connector, any organization (§14.9.35 GR11/GR12).</summary>
     public static string RewriteShared(string name, string image, int length, FileRecordLock phrase,
-        FileRetryKind retryKind, int retryAmount)
+        FileRetryKind retryKind, long retryAmount)
         => _reg.RewriteShared(name, image, length, phrase, retryKind, retryAmount);
 
     /// <summary>Governed DELETE RECORD for a sharing-active connector (§14.9.10 GR6/GR7).</summary>
-    public static string DeleteShared(string name, string keyedRecordImage, FileRetryKind retryKind, int retryAmount)
+    public static string DeleteShared(string name, string keyedRecordImage, FileRetryKind retryKind, long retryAmount)
         => _reg.DeleteShared(name, keyedRecordImage, retryKind, retryAmount);
 
     /// <summary>DELETE FILE with a RETRY phrase (§14.9.10 GR15 — the '62' file-sharing conflict re-attempt) and
     /// the GR18 OVERRIDE flag.</summary>
-    public static string DeleteFile(string name, FileRetryKind retryKind, int retryAmount, bool overridden = false)
+    public static string DeleteFile(string name, FileRetryKind retryKind, long retryAmount, bool overridden = false)
         => _reg.DeleteFile(name, retryKind, retryAmount, overridden);
 
     /// <summary>UNLOCK file [RECORD[S]] (§14.9.47 GR1).</summary>
@@ -297,7 +297,7 @@ public static class CobolFile
     public static void ReleaseSingle(string name, string recId) => _reg.ReleaseSingle(name, recId);
 
     /// <summary>Evaluate an attempt under the RETRY discipline (§14.7.9).</summary>
-    public static string RetryLoop(Func<string> attempt, FileRetryKind kind, int amount)
+    public static string RetryLoop(Func<string> attempt, FileRetryKind kind, long amount)
         => FileRegistry.RetryLoop(attempt, kind, amount);
 
     /// <summary>ISO §14.9.27.4 Table 19 — is an OPEN request unsuccessful against ONE connector already open on
